backend/routes/requirements.txt
Flask==3.1.0
mysql-connector-python>=8.0,<9.0